# refine decimal multiply, avoid cast to wider type