--- a/datafusion/physical-expr/src/expressions/binary/kernels_arrow.rs
+++ b/datafusion/physical-expr/src/expressions/binary/kernels_arrow.rs
@@ -35,9 +35,7 @@ use arrow_schema::DataType;
 use datafusion_common::cast::{as_date32_array, as_date64_array, as_decimal128_array};
 use datafusion_common::scalar::{date32_add, date64_add};
 use datafusion_common::{DataFusionError, Result, ScalarValue};
-use datafusion_expr::type_coercion::binary::decimal_op_mathematics_type;
 use datafusion_expr::ColumnarValue;
-use datafusion_expr::Operator;
 use std::cmp::min;
 use std::sync::Arc;
 
@@ -686,7 +684,9 @@ pub(crate) fn modulus_decimal_dyn_scalar(
 
 #[cfg(test)]
 mod tests {
+    use datafusion_expr::Operator;
     use super::*;
+    use datafusion_expr::type_coercion::binary::decimal_op_mathematics_type;
 
     fn create_decimal_array(
         array: &[Option<i128>],