@mingmwang 2023-05-11
And existing UTs show the result is correct.


```
query T
select arrow_typeof(c1*c5) from decimal_simple limit 1;
----
Decimal128(23, 13)


query R rowsort
select c1*c5 from decimal_simple;
----
0.00000000014
0.00000000033
0.00000000038
0.0000000005
0.00000000096
0.00000000105
0.0000000016
0.0000000016
0.00000000165
0.00000000176
0.00000000176
0.0000000026
0.0000000034
0.0000000039
0.000000005

```