@mingmwang 2023-05-12
> Good catch. This fixed the tpch-q1 performance. But for the root cause, we still need to improve the i256 operations. Otherwise for queries containing decimal array to array multiplying, it still suffers performance issue.

Yes, you are right. We still need to address the i256 operations performance, otherwise when it runs into the related logic, the performance is still not good.