@viirya 2023-05-28
FYI, I'm going to improve the i256 operations at the upstream: https://github.com/apache/arrow-rs/pull/4303.